Move `get_state_transitions` logic to `GeneralizedDatabase`
